refactor(Cars test): 테스트의 목적에 해당되는 부분만 테스트하도록 변경
