+++ src/test/java/racingcar/domain/CarsTest.java
@@ -1,12 +1,11 @@
 package racingcar.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatCode;
 
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.infrastructure.FakeMovableNumberGenerator;
-import racingcar.infrastructure.FakeNotMovableNumberGenerator;
 
 class CarsTest {
 
@@ -34,14 +33,13 @@ class CarsTest {
     void findWinners() {
         // given
         Cars cars = new Cars(List.of(
-                new Car("name1", new FakeMovableNumberGenerator()),
-                new Car("name2", new FakeNotMovableNumberGenerator()),
-                new Car("name3", new FakeMovableNumberGenerator())
+                new Car("name1", 1, new FakeMovableNumberGenerator()),
+                new Car("name2", 0, new FakeNotMovableNumberGenerator()),
+                new Car("name3", 1, new FakeMovableNumberGenerator())
         ));
 
         // when
-        Cars movedCars = cars.moveForward();
-        Cars winners = movedCars.findWinners();
+        Cars winners = cars.findWinners();
 
         // then
         assertThat(winners.cars())
@@ -52,15 +50,10 @@ class CarsTest {
     @Test
     @DisplayName("자동차들에게 NumberGenerator를 주입하여 생성할 수 있다.")
     void createCarsWithGenerator() {
-        // given
-        Cars cars = Cars.createCarsWithGenerator(List.of("name1", "name2", "name3"), new FakeMovableNumberGenerator());
-
-        // when
-        Cars movedCars = cars.moveForward();
-
-        // then
-        assertThat(movedCars.cars())
-                .extracting(Car::getPosition)
-                .containsExactly(1, 1, 1);
+        // expect
+        assertThatCode(() -> Cars.createCarsWithGenerator(
+                List.of("name1", "name2", "name3"),
+                new FakeMovableNumberGenerator())
+        ).doesNotThrowAnyException();
     }
 }
