+++ src/test/java/racingcar/domain/FakeNotMovableNumberGenerator.java
@@ -1,6 +1,4 @@
-package racingcar.infrastructure;
-
-import racingcar.domain.NumberGenerator;
+package racingcar.domain;
 
 public class FakeNotMovableNumberGenerator implements NumberGenerator {
